requirements.txt
behave
discord
python-dotenv
PyHamcrest
requests